+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/ca.crt
@@ -1,0 +1,33 @@
+-----BEGIN CERTIFICATE-----
+MIIFuTCCA6GgAwIBAgIUNkPn1B+YeHUv8D5QvrkCnTF1T5gwDQYJKoZIhvcNAQEL
+BQAwbDELMAkGA1UEBhMCWFgxDjAMBgNVBAgMBVN0YXRlMQ0wCwYDVQQHDARDaXR5
+MRowGAYDVQQKDBFUZXN0IE9yZ2FuaXphdGlvbjELMAkGA1UECwwCSVQxFTATBgNV
+BAMMDFRlc3QgUm9vdCBDQTAeFw0yNDExMjQxMzQwMjBaFw0zNDExMjIxMzQwMjBa
+MGwxCzAJBgNVBAYTAlhYMQ4wDAYDVQQIDAVTdGF0ZTENMAsGA1UEBwwEQ2l0eTEa
+MBgGA1UECgwRVGVzdCBPcmdhbml6YXRpb24xCzAJBgNVBAsMAklUMRUwEwYDVQQD
+DAxUZXN0IFJvb3QgQ0EwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQCz
+4ZPsMrElTYPIrlcSddMqVULoTlDCINShnLLk6zm80d3XQQbZklg7MjP4PTjzQSR6
+Ko4Pz/Wom1VxelNDMCmYT5+IP6xEynZWFj7rtXghHoXazxfxAzF62Y10qBaNViL2
+JpVNWTY/mCSLyzH4x2hFLMUY7QIsE7HBmghsJgDU0DkeqSVXjNoF0VCsEv1lXgxI
+75jQ46ZTu54n08xWFdd61kOtPq2VAojuoqEDFrMM9L7NdsnaTcfv+Hybj50CEHmV
+QQwI0iHyc9vqD5QMKtMiynuxke2+Z25+bBgOpmGnDfSza25Z4md7t/Yb5zDE5sem
+5P8TCCoQF9cNxjgk3OyseDSthkIxgSfAzDUgfnMBWXXg+k4MtyJjkZuIb1t7rczF
+fMOVB3ynk2XCDcSrhTB1pjfKAGw2Vt6MmbmYYajS3Pifu23V118YWkWObRH9W3Fd
+WdsGa/HCSLiOuU88KvA+nxNyAzFYTWCKSwNaHl1Vb8tkTkfJil0TRGP/S7sSmR24
+Grk8pVOxjRNUrcTdi3k5l6lsbuv6pCNdN55uJNoKMtP8fzg8OnqGjXZ/dnHSCdJS
+F9CejDZNDGBwdBhd2qN5tdC3xsMAKM6tXGdnOwhS/aEhz1E22nCabwGJVf9n/Hyz
+8tIdIxD5HYA7F3fxbJwWhZcu+e9oVaMdA8Aa59LWwQIDAQABo1MwUTAdBgNVHQ4E
+FgQUKfldS5qiKJ2Yni7TiEWhHxR91PswHwYDVR0jBBgwFoAUKfldS5qiKJ2Yni7T
+iEWhHxR91PswDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsFAAOCAgEABuF6
+Bfdys9J8IizsUh9kCTiTx3QKOOEiVXYT+SzmWXhoKSBfcbee0lgs3y+0KVN4Ry0d
+rk2IdkgouWAP9B/epxCMpdKVTINoxasYT6rDCouCdP+fG5sbHsOKSz8fIh7/b0aP
+en5AfDVueop7DG8kQZHmBbIBySEBi18jQaE80PgelZuCf2LGsEdBEWaRwTiMUXZ/
+vtG7o9CtaKCkD0Y5TwIpER6TsKWR40XDk6dL7gzcB8qwKmWeCKe/7gre++I9NyT5
+x/utCzer6CERJL4thdi0MDhOPl10eCo82mH9aeLNbX7uTx/9eWQ1KxcdvBPOTTDR
+7LWtFN9vDcBvROIi1K5uYkWEEd9uJjxq+8ox9oDJI5pvFQ2NQne8aoitxFQtt+VH
+STjyWudxUjOyIUAp71qqQOdSs7t0R55l5rbBJsPlFL17uRkKPRYtXjiD370Q7dfd
+bTRPFllEQRO41W3RmfO+aaKkv+/zFV+s2ERGIk9zNpYXjxPtG9imVKE2aUGW7quY
+9ZDFv6yDQ5lekonVigW5tDBrRzYbFR4CF7PlzveMRDqSFLkOg2kG/TXAG5YCxxO9
+daDA0nPLy69OuA5mgecnZPzPNRGhmy4kzFiOXR0tvs1n0UZZOZpJCNgIMG5jR44h
+XL+HJs9105c5NrutgfiXmdfUXkycYTdpTyrxA3c=
+-----END CERTIFICATE-----
