+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/ca.key
@@ -1,0 +1,52 @@
+-----BEGIN PRIVATE KEY-----
+MIIJQQIBADANBgkqhkiG9w0BAQEFAASCCSswggknAgEAAoICAQCz4ZPsMrElTYPI
+rlcSddMqVULoTlDCINShnLLk6zm80d3XQQbZklg7MjP4PTjzQSR6Ko4Pz/Wom1Vx
+elNDMCmYT5+IP6xEynZWFj7rtXghHoXazxfxAzF62Y10qBaNViL2JpVNWTY/mCSL
+yzH4x2hFLMUY7QIsE7HBmghsJgDU0DkeqSVXjNoF0VCsEv1lXgxI75jQ46ZTu54n
+08xWFdd61kOtPq2VAojuoqEDFrMM9L7NdsnaTcfv+Hybj50CEHmVQQwI0iHyc9vq
+D5QMKtMiynuxke2+Z25+bBgOpmGnDfSza25Z4md7t/Yb5zDE5sem5P8TCCoQF9cN
+xjgk3OyseDSthkIxgSfAzDUgfnMBWXXg+k4MtyJjkZuIb1t7rczFfMOVB3ynk2XC
+DcSrhTB1pjfKAGw2Vt6MmbmYYajS3Pifu23V118YWkWObRH9W3FdWdsGa/HCSLiO
+uU88KvA+nxNyAzFYTWCKSwNaHl1Vb8tkTkfJil0TRGP/S7sSmR24Grk8pVOxjRNU
+rcTdi3k5l6lsbuv6pCNdN55uJNoKMtP8fzg8OnqGjXZ/dnHSCdJSF9CejDZNDGBw
+dBhd2qN5tdC3xsMAKM6tXGdnOwhS/aEhz1E22nCabwGJVf9n/Hyz8tIdIxD5HYA7
+F3fxbJwWhZcu+e9oVaMdA8Aa59LWwQIDAQABAoICABPAAaVqQu4vMrgY4T1JA8aG
+/j3CpIYOA/6bFX7CL3CSg9aXoIyIr4iPWGqp8SvvM/W4YKjuisDIjlLzbPsoaXa8
+Tk4vvIbAionk7yNi5TXQs75spMqs8LS1+HytnG1H/vZzEctAb27/0cpqBNp1bXbD
+5b+Y/Tcw8eBJfOF/8jErjAv4Yx3X3JPNMbv06U7FCtkCuo774cYZiFHD5NByXTpy
+1lMqGbWtxbJa144e4smaQS1OvLar+db4m9jaJbFldO386nj49mp4C1Ypf/YuiQhN
+vlNGrBlu4pA9ps4L+VaV1w9QjTkP/U/1iK9zXLTPyG3EAJgaukCSV4W/BOjSO2d8
+rn/AQok6WkQO+swcTbRA3avsXaRvDh6e/P+3OyZnsVRGYUKkTraq6mYd1o1RB7BN
+3DrKFK+vI8XQEOsyV/HZm8kF+hguQ7XpTkMsd2uGScrQzyaKKE2WdNvkab7tjECL
+UkdLw+G8VqFOQN18xEJpnE5CM9qMKpPnNjd+oo80oTtuosEWGL9t95m12psH3Ii+
+FSLjmZaSEs+JTK0+CDAjgxAqjzNc+hVx11ln6PXt8QHNuxzbel9VtetZwbAteWNx
+F7y3zkqYowCOuwFZCs3qEoOiI66nf7/G4SdVxcWMzyd50lEXhzEwTD9zYttyZ46A
+3VRw8T0M0HMY/OSFMyHZAoIBAQDM7ks1kYOM9/LE6JFMDqrjbDnjAJBOBDRki0iK
+tGvC64tGmQlEgf4AyY51QNTAkiuQc3UJv/VfNz+E+aJx6K57z5BQt9ZfLT1jtIUn
+8/r/bTwIpVsBGOxoYEtr91WoQX2ngmrQqd/Az9gfd09Il1b4bbisAvlfKM6V5Bb2
+z8cz4Zf8dqRjDyHA06lqopK2dAJIrJEtu+4ooRs5aSI0XgklsQvCuP8tWBJ+WZPr
+LrYSi10a+2eTF3x5IO0CkRpVREroDXqQk1lvJyGEBJDXN20plftOPHkhh+/Uj1nV
+dMLwthsvhvfsm3wxA0WSTIMzu3MF5uikt8ohfS+XrxeF3QiZAoIBAQDgtTkBXXXi
+PP+J69yhpNdxkiycX1AVVO5ojHbntCLqtz9am/ulHgSh3/Yp0Khr4ThsdVmFDbeH
+srkagU8PtQj/HkRJ3VPMACphNReZdV6JdGdPy/GuI0/ewS0tQblFxcX8Op7Nsnhn
+RG+FOgBu1LrcZbYKNmCyJNj5i6CYem59c6jC8LOYqNgMJO0qh62SwpFBv5kYTSrE
+JxnITRrspiOiVMFjO/OS3LdLLdQIFuNZLKpmh/uB0UbJSihhwwkmnllAoEcmRh8q
+V3fuei6kT9/OimA+tZ+81a0aO5o/5gSE15H7WSp0/TLDdgRh2lDrZQ0hVb8jimpM
+lvWRLH3pXtBpAoIBAErmWWS1z3/pa7ieS+Shg8HlX27hWNy62q9Y3OWdNbmjP19J
+NcLtQ/njsQaX8GqFwhhkqoUqvLLVTYj7eGzrn0IvRc7KsjZNIZ/cbcKBgN50fTmn
+vhAznMw1TbujRZJehV+A4itruxUMojA1+itM7QOVjqaH2770+jR8jgwdOvByZotm
+jYSLFkoU0lngG4FF+0O2lsVBZ7U/1ORx0YeUSwGUqPi9hr567x9w/Ag8JV23xArF
+kkCHl/3pk9aZVCeP81THAGHxCQIAKlbHGG0gm+xUHeqlXZxsw7JLxoimyU0TMfJ4
+AQtgdM6xJG9pOZbhzxN8JtF1PxJoX8mvvAg4FhkCggEANfmRzdAjRsqcHwb2oiU0
+UuVnrc//LoCOG7LRxOoyEOsfXqZlFc5WEoP3XyftrKwWn4a4kAsxcHjvmXxB/2Xe
+2Vjt0khFnpTTmTHlj0jwAlhhkHtOfCOjRcBlF5/X3viYi/OQvZwG69c2amv6OqUC
+lXog9ecF5wSO0bJGSxntfFjRyrGCkkLMUm50QWkBumTd+0sIfRvcD+3MdZ5/vyRR
+Hub45Ns7sKUvqfFCIwdMaNkfvS0grpgURrl3ZbQxeINEsQ9Ox9TEG6mS90eF+JXv
+Dm+Smr2JtxQkVk7GyF5562bUe+l5F0zSarzoZJDKKyDEvKMaHHFdAzSnm81yGSJb
+MQKCAQA3z0zVGFQM7HY1mGU/GvBOiqVbnOdEaKOPBI/p4hYpUin+jyTzdO3IkS8i
+p9HQ/PLuknusjtw1tGg9AFGD/LFOpMMvAwTETKFSK33KwAH/S4uUwHoYusGOotiN
+6cXmbt4UlGWNxNUcFCoi8xyD/Rh7nmCbV17Nt7fHTcYe7Wg7HEBYLPNwkIEAC7MP
+5uUFq4VKuvU9tG3MxvSRBh8DJYWZgGkg2WUXyxH2UgH0F8WUJHpxD0sDURYJTsGp
+eXFCMaRGQMXHT17WyRy9pzZgPm+hYqLuGEzpMCgtKQWMmaDXDEgWvfJ3ofZdlpqD
+BeloSKvtuFuawjgyuSNzh/0jvSp4
+-----END PRIVATE KEY-----
