+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/client.crt
@@ -1,0 +1,34 @@
+-----BEGIN CERTIFICATE-----
+MIIF2jCCA8KgAwIBAgIUTWq6OddIP75zvzRaygwD/Cgrk/4wDQYJKoZIhvcNAQEL
+BQAwbDELMAkGA1UEBhMCWFgxDjAMBgNVBAgMBVN0YXRlMQ0wCwYDVQQHDARDaXR5
+MRowGAYDVQQKDBFUZXN0IE9yZ2FuaXphdGlvbjELMAkGA1UECwwCSVQxFTATBgNV
+BAMMDFRlc3QgUm9vdCBDQTAeFw0yNDExMjQxMzQwMjNaFw0zNDExMjIxMzQwMjNa
+MGsxCzAJBgNVBAYTAlhYMQ4wDAYDVQQIDAVTdGF0ZTENMAsGA1UEBwwEQ2l0eTEa
+MBgGA1UECgwRVGVzdCBPcmdhbml6YXRpb24xCzAJBgNVBAsMAklUMRQwEgYDVQQD
+DAt0ZXN0LWNsaWVudDCCAiIwDQYJKoZIhvcNAQEBBQADggIPADCCAgoCggIBAKsf
+JKXY6uJhYEXvia9n818MJZOLkysZxFKa551dFWI5pzQNtswlEp5bhsk9L5v3nVik
+LJuYfnkQqBxyru9utiDMJvfNHR/LUM0M4aA0jyvnpPI6Zu09YL49kz0O90Ldms0V
+DGhj2iYG2W6Br0jr5grWMXWtiJSzryjjL2Tly80K82Wf/Gzb8XL1xYNb6R+3v0a8
+J+QzgOERNtSiH6hfPoVk/c/oLv4PusliDmUrcPwnjw/coJb8mx0v3tVjht8rz8Ca
+T03TUPHjlawJRvMc2ajcpHV398xJHN7VBrsjSlHGbgUsGulB02zl6OtBIiqzfQKN
+0kUsw5aaZ0Vv+yHtrZHQ/T3bdSdyjvgeNEX1Imy12vjy5P88ZhI32CiEHA64U/SD
++WnWnLrg5tj8urOEJQxqtygSs5Op/DIC3w7UgsoJ4NjJGkIn8CG7sfqV2NR8tdDi
+LmffTNJnzns3llv/CDqXL5T+m7181OTsoAuvBXcBKhwAUIoPqu5QM/kmzwwsqJSc
+RZtEGy9NuSYsrMpnJtTIxpFsOnOlPthNgpFJ0yi3BKujYz5sEzeswL+nAhsdbxm2
+NC5APVBYpR85GjyJSmRMa58dyWtWWAotOFS2Fz8i6gJnOUktZWBXUZJYz+g4KdfK
+Hd0zKg0CBGJDDRBUq3QULaqXhEfmDLioBoc8PaQ1AgMBAAGjdTBzMAwGA1UdEwEB
+/wQCMAAwDgYDVR0PAQH/BAQDAgWgMBMGA1UdJQQMMAoGCCsGAQUFBwMCMB0GA1Ud
+DgQWBBTSPYdbNviiTxjswQwJg5Dd3w2dPTAfBgNVHSMEGDAWgBQp+V1LmqIonZie
+LtOIRaEfFH3U+zANBgkqhkiG9w0BAQsFAAOCAgEAompovkmzsQCOwHNVN48ZT+R3
+NJiRSdkSDSmNJS8zmq5xt0XnG/nhZkwy3zdiILOYPgnlBTdREDV5UZ9QI/K+Wk0m
+7iKVzqgtJS0zOc3TMCElQ7oQiGU+s9oLtAE4QhIph89Kh61LAGw4uTZLEVYysG7F
+/q8GSbL3G5ZGeE9olo50+TZ1xEaXyQNtzwLluOeggkXf55SLyKdBWT5cZZsH5l0b
+cFv8C3hDpHvb2F1f7Ey2jljW2rv7xcYeXfdJQbSLzZotfDyEcW0VMchbL0mPVGhj
+6PAwGPOXhfsG+Ztzr6taL8C3VAXnHd+HMeXTqE08rp0IGYzo/3ix3to7cq4P3Jme
+2F0Bbwa6tH2QNphzBHt2rLqJ/yBkiX2nSLJECPaFS219ceJn9vVIp3OYrdS1JU7I
+PfQ6fyootYfwDozcjGzD2AcvMdrrak69Y7o0O5dZw6AcTWiiXsjjmTngwmnhfSJI
+KzLH1Fnz5IMyoHiOZTAVu+pMPTV576U7oc9I1j7wrzm0u2SZ5N7sfJ8bfi0h0MJl
+dQsZTBL4qREILeNcReLEsmLX2b0V7ocYxzTjXSqPSd+AWHjkOC+4MlG7jIc5Clb6
+VZZtA64jhAVhxM+aCDm0L7kpkowf9RsXK15mkkrDUuVBlGVCL7lpqygm0RBYDxFx
+V4sSOJDOWMpU6GSPFwg=
+-----END CERTIFICATE-----
