+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/client.key
@@ -1,0 +1,52 @@
+-----BEGIN PRIVATE KEY-----
+MIIJQwIBADANBgkqhkiG9w0BAQEFAASCCS0wggkpAgEAAoICAQCrHySl2OriYWBF
+74mvZ/NfDCWTi5MrGcRSmuedXRViOac0DbbMJRKeW4bJPS+b951YpCybmH55EKgc
+cq7vbrYgzCb3zR0fy1DNDOGgNI8r56TyOmbtPWC+PZM9DvdC3ZrNFQxoY9omBtlu
+ga9I6+YK1jF1rYiUs68o4y9k5cvNCvNln/xs2/Fy9cWDW+kft79GvCfkM4DhETbU
+oh+oXz6FZP3P6C7+D7rJYg5lK3D8J48P3KCW/JsdL97VY4bfK8/Amk9N01Dx45Ws
+CUbzHNmo3KR1d/fMSRze1Qa7I0pRxm4FLBrpQdNs5ejrQSIqs30CjdJFLMOWmmdF
+b/sh7a2R0P0923Unco74HjRF9SJstdr48uT/PGYSN9gohBwOuFP0g/lp1py64ObY
+/LqzhCUMarcoErOTqfwyAt8O1ILKCeDYyRpCJ/Ahu7H6ldjUfLXQ4i5n30zSZ857
+N5Zb/wg6ly+U/pu9fNTk7KALrwV3ASocAFCKD6ruUDP5Js8MLKiUnEWbRBsvTbkm
+LKzKZybUyMaRbDpzpT7YTYKRSdMotwSro2M+bBM3rMC/pwIbHW8ZtjQuQD1QWKUf
+ORo8iUpkTGufHclrVlgKLThUthc/IuoCZzlJLWVgV1GSWM/oOCnXyh3dMyoNAgRi
+Qw0QVKt0FC2ql4RH5gy4qAaHPD2kNQIDAQABAoICADZyg9IjSWU7b4sBExGVJTou
+XrxqEisKGU3hTOXkqruktX9aB7o/54Msgx//0GizakS8hBcD9+Xkd2EJ2oCDfP7F
+hnadNHGb2vtLaHT1lEJ3g6OxEnPU7jXJ2NH6M5F6ellbo6WJc+1ai49eQxwUNw9m
+pLkZPJZ377k2RCu59FUq8VP6CjLG7rXR3oUT1wLJD9rdXmH9QzoZpJUWerUzjOyb
+vCiNa3paHegaw5LMNjh9C8eTvyMbf+1TAHdsIgtbYNCbKLWOEbjNXZ+uGf2bxeO1
+RLy9rcI51y5NoMUlju2jEfmiVXO3hqKlKq3D/S+SCaoo3uk0u4ztVwOghEsA2ZNB
+NRWEMiohaU9BNXZMt0TUAEXvooR/ig8MCKEQMjXpglI5T/d9VAdmKYakA3HOaiHY
+ph6zKNtKgSeFcn7ePZ/pZiMFhVHXQv4OfvpN35msdkXxbuNRu3CL8mON8z3IQoBl
+nxfvFWy9BQ1Qv8WaZu5XQerC1NT+AfOyy9ShHL3ShSalJNB0sbJ93LrFIQ5tOHPH
+97A2yZlezleozLokkx2c/GazPsjNQvZaqBs1aVa/C1r8MaIS1AyXcKTYdEGL+rZQ
+JG5VkzT502p38lBBhXh0CNZex8HOUl+tGtaZXCbhhDhhNzFiLsSceqk3cRr4VmYJ
+x4/SaLcpN4mU9NhV4JnnAoIBAQDTm0DlvH9OeuQzWYThCjw8k+qL7HlEsF83b7ST
+koJEMBAkLcKmtn9i6N1P2bRYTn1jsRtYjzux0I6mZF6ij8kAhp9sTcmCtLNRHJb6
+yRn0Xh0YcgP9sWR7EkkoRwL6H/tGHdk8nbA81SIxvf1u9HA6CCAfMONfwEhO/nfZ
+GxZbqLr2tVzDBKLYEEyKbGp1jh481SQRucoNUvzNpwEqHJQU1cpsHjYKaKQsO8iu
+nT1D10U48dxpW2HB3BetJ8Vel2bY1BgQmTOTmSdZlVj9uRSrFd8ObySFVfcDlyjP
+DwEQwfb5n6UK5IEl5jCiT5G2Ef/g+vTXbh0J/lIyZtEVpPRrAoIBAQDPBZKN4Gee
+qX1Jw1epkZS0HlqGawuAKiYbWmzxq2Z8Okm/WVKSit3JsPL+xQU2jnLBwkWaBoF1
+0r4AHYES/++JKiSxVor98heuer0iqjr7pWP0ytlZdiEcXQhbmZNzPAccxtnaGXSe
+24vhUG5jEre50MRiuLaOiXOFFcO7xXe09bIAau8cVmEN+VQVWQNxBiqASWZZlxvG
+K9NbtdZeE2C+qjcjhdQgCbaM8aPIejnjzlm74Y67eSnfk70qw3Or0ogNGlxVq2VA
+bHHA+Pw5vjzQyDfzq8UB+Pte+sbKHgGrZ1bbEYt0zxW0rowOjSf6NbadYlxaCsrN
+JM0yeKADXPHfAoIBAEarP+Y5qI6OLi4UBHOflnRotdkbFXjiKjp6XOU39bdOEeXz
+PmGs6ygeOuMTAs7HD8s+d7KAy1/cFfft50kFGgzJ6O4joFu9e3FoSefhR4C7+Z2T
+Xk/YcYVEmxgPdygPjDxjdYxv4U6/o1bXa48lpWFMG2YvTpBIyB1RkLb5++KKnD4O
+Xbdnkfb0ootMH9W4q9KHIC+nS5pdj74iuxnUR4QZziuDf2rEfaekI6ZlLvO8LgsR
+gxnQgfpbRbD3akvn5lodoxfQAlCVwnWBkiHSMNveRwssBRdemUsirWdRTHrLPG8r
+aNQVcig1EyLxzejvY/Q/+4sC5EHVJKBehA9lcpcCggEBALvl5+XD5ui4FxJLLZan
+ntt4rVtRbyfNmba79SPIGQXp9aOqU+rUBYeZhoO8mQWydu/r/zZUb05cj70VCmEP
+UP8i9BjNXuwsvlfL6Ty8WzqqG7ZEItAzCsxumDZjhtRRiHcS+Qoejp3c0PIZCYf8
+pisOFCEyW+e1U9dYMMaN7R+e3QfKBuqmcAK56RpTJ29/VDYJ6M6p+zh78KNNwJq0
+ZdvILLGeNzE4/6qfa7N7VG2yjfmcOOtphpr5wfk9WxB5zqht456SKR03EGxA0NUs
+GJC7GSWeIvquXKERZjpuz78d1VSScQv9JUtQHajbb4yHEg95Bi9nmjhmDaSgnUNQ
+KesCggEBAId8UZCh8kzIWe+Z2HxHb/I08APXGAO4iR++I+/lX1Ayq2OTWapSBpZ2
+upzouDll3NsB9jlZTjU7yVzGmKgPRR5jUMSKs2WIsYW/9n9VBXYRgKXvrwoIEppJ
+0Xp+/8bQcDecEe1nDbfpNPxQKySsnzlayPfBD68sv8/1hxNuByaq35XMxiUsuXO7
+VHfRkl/FElG5XRhEE0jDBF7c88dYP8zyGSjzb7DQETGXkj41m5z+cnzYRFkRbaLN
+WAidV+DOoaitVclBrSHGFq8ycaP9cPkGPN1F55LuZUYT7RL7VjMNb3hHwPNGmKK7
+m3+Z6a2QeJ8jJcxzEfKY8rK01d9HoSA=
+-----END PRIVATE KEY-----
