+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/nbs.key
@@ -1,0 +1,52 @@
+-----BEGIN PRIVATE KEY-----
+MIIJQQIBADANBgkqhkiG9w0BAQEFAASCCSswggknAgEAAoICAQChLM09c5EWjHn1
+scbb7leueqqADVtJeo9ko4rCbMWUOEYdS6nnRpoKvI+SG+36sEmTUBE4WASNKXWA
+p5VPmvzeFYvPIuJJNfr2bFS2BNaFUtk8Kj0G1fJ9a4JJ5LOGokS24F6uaEK5ZqW0
+TkYkPHQ43GAqN2pq3q681sdfAaItSqJg3Aa41YmWdYKxgPWVV9ADsmv/zm+zWsDK
+rOPET9GFrdWQ0YPcL7QW/Gw373rZadAdIPvHXtzFXYUGO5V81+7DaNAHGta9YL9s
+Bo0cu4tYgobI8l/V8uvZXvrQLl6GbtwUGYTgfUxqmBmLUcw4jwPWmvrKTlI3dG4Y
+oB4KfD0bZN92xhExoBDrtkMCeko3FBnxetL0fWdxPZ811UaAwBCX0BJZQPUKPnKn
+qI3L/inFM/T+nLg2h0UUTyFQvjyko+YEbZAscDeh3kLyGIUwdVwpoU5vwteuD8/e
+0GEX6vz722ufDp3rvLD0b6I8M13WA2dKhqRM5g+Ej8n2S2hUxuKtDIodYEtcLkib
+T2RyccWcu4+JRpYqCM0w6SL9ModW5ZHFaTKweMlloqpAH5OhHwtyklFyVCfZ9rLf
+O4/Kx61U4m5stGmbj/dCBhQ7azqAVugbgIf0hRtfBZ02BedIe3iUGwJV33j5v2dN
+uEt3aTQMcleo9dPYyFmHzXulaCBNmwIDAQABAoICADZAOc9K9JXF8BFS5zBFQC9D
+K9vRYx5PCjAiWkUKTtEOxg77VnO6EIS3po3UJ9z/dKyIwVsZSWLohxwojHoQnbf1
+QauVvtxMtq42lab05u5WjjeRshv4bNhplR/hQsp5GS8DQ+tpvbfhRLp1tGdjD+3J
+w/bB02iojuGSW+ZvStL68cAvn3WcruS9piGwWjmMJsDv0Cl5J6ET1ZwC6OG9yjU4
+Z5AI7dzyE6i/C4AidFP1KW98mDMqC57FVU7dRHqSxlPxFkCl+I0jIRuv8tSujf9b
+ZBoEd1yxGVEABuWHDhsq+porYs4bbtrcWpDKFjcdcHnOjccF6OewoZRZZPM7kWd7
+wFgOZeyGh8uialnD7tKBfwbLDDoH1ZMz16UT0oINUQLmApEHYfw4KKcNNJ5X5SPe
+RMjQR7weHlzxwPPm9UuA3pwtnENPBp2ujKSfLumt/UTI07pvysHYbc6IMJHEPw2m
+q/48U/gK9HKrWhtoiMG7jc+hxhCxGSXbY1sxUuV2oC5CoCGAgUvAq0DLKWuTkQxM
+dg04uYecKUyjL7A7loid6L/h9bi6GjwTAR9PyiTle/dBx1W1au7eKVpJMFTLCsFR
+m5IKDHIaEKquREe0GX1t00RowWIHzWr5CxnCvbf0dsNJooqUdQXTC3xScx7k1DQ3
+nfhdF03pmtuK/D16/wBRAoIBAQDCIfXy7J105puFwj25+mSwumXcyNU2IGzKfLdp
+zNPqxytTrAuDiwruMLEIaCmRk3ABc+W4Q4VIFi9JgYgfkirkOgBdIhvP0Usd0Cl7
+cbW14qJcwCt7vzyCEMNZFi4axkcpSF2bYm4XB0A3PNoeVi9MzVT0NIsj28PWL/5v
+XyhfSW2Pb5k60FYPVyJ3A+rQ0avFkiBvVa4iQlvdz6VCvadayI8FLNb6i2mrnZek
+Xpx7eWJckIoYhxi3QdIBob3aEGM6Fir307iPA/tT6JiNrjJNfemVA7PnicL1Jz9I
+IV+RkLcfTq3KJi2CQfw0+ZjYUMScTEx+tKauDAiX5gylkJRxAoIBAQDUigmRuFWQ
+63hqaXmrZgvLV4LpqcPgfausMUAf64tsqVGqpVdpydX5k2YcRZmNUD9OOJOnqCNv
+iRlwt1EW7Q9A9LP+Omx6AukrejECF1nxGqAhXonGHZJFksnQO2RmGw1oURitmOUc
+t+4kn61+1htCKwxf895LGKlaJ2E1GOIde1Tp6cGjvUAicqi6dPsWOEFpjFWVwB+N
+PjlqdZoyqI6Ss77kTxGrSA+k8SOZeTB8ozCA+Fsbj87sVfdcrfxFTLnnL96oZJw5
+jmcgeemAIa2POsNCrZ7uzGWyB731wXf40MojTkXjRpkDgYzuY1M5wC5KMkjVtM6a
+S/Ok1QqhkBjLAoIBAG+rHVDJLqGJ7Tf5zYvnLnhCIEnp9EsYR7wKFW70dYn9Wzzc
+IBtgpb7xn0PYtuCYULtlMWTW4hnfOQAtVGHksGtrub3Y2uql13GkxdfORtoDLaXw
+nJKoJGQkxFnAWnDoURK8H6w76FuyvguII2eD591j11RJzSepSHMyVEc4E38MO8tg
+RGICaR6V3qtbQ30uvoH9Vk70mB/4z+M9iDBl65ha+XnemyKvgu9GzW2tRpH39s6M
+DgOh4mO3ptdtdEm1V36YNcssrGMHxVA04VvOMf1FSlF/OcMcCVr/YnVZTSHsfomx
+tmIrp83fqvXGBVeYzh+h+/Z+b8p5/Y9ysx+hlOECggEAEWH4xmI9HU0k3A8kVFBU
+Rp0xp8qqqoGcuMiOZg3tqUVv2HUNYSmQxoi3oThQBq65frxB10R5IXYlDoLvruBU
+jwt0bSZ6GMCEMXoc5/7g7EcxMYl52yov/GC1AMLawz8TKaT4T7phgqfIYlZs6iK4
+uACtUmuvkpTBxQ3TdXRDigW8PCVSIQIpbKfUBOvsXTTcpLVFiqnb6sFcEMjtg5cl
+kfoBcsrEty19mg/eyX6P6KGwJdwqF2b7lXiycnIIX+KWkcHog/02ib6L0Kwn3MLk
+XwJ/VcS9aupsptLbRm48Ua1B2wFb1erAOPkvfY/CDscZAhh6vH+9JgFwO9rws6YF
+YQKCAQAe1fAXOtEmjuYbeWBConZqgPYPJNfjyd6KSh+sSVbQRyTHvmTwpsMtFCJv
++FYrPziwTa62/hC4njp4IuFH9YxaJmhGQ7j1ER/SRo0pDAG7m3PxV/pOpNudlTRq
+5mUQ2H06HqAOEyRwAoIW1MdDSC9ijvslDE4Ptg4HAhPFyQVn1x3dtgdxZ1Qoko6V
+DUNw5BiG1pplLQsXrOMS07o5hkrUtepSMeCXLjdLl+G0dpxbZFue5hR4k5fdjzeq
+BeQZKFC/JAZevsnT5gsHqg0REqRaltphB0Jh3q7uVYIYI1iB5r6UtcDG8ChMav4S
+VI683bkEAhP7S4MCMp3AkKXH0FVd
+-----END PRIVATE KEY-----
