+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/server.crt
@@ -1,0 +1,34 @@
+-----BEGIN CERTIFICATE-----
+MIIF+zCCA+OgAwIBAgIUC6vuD61j9gr3fW3v7Mf66yu2myMwDQYJKoZIhvcNAQEL
+BQAwbDELMAkGA1UEBhMCWFgxDjAMBgNVBAgMBVN0YXRlMQ0wCwYDVQQHDARDaXR5
+MRowGAYDVQQKDBFUZXN0IE9yZ2FuaXphdGlvbjELMAkGA1UECwwCSVQxFTATBgNV
+BAMMDFRlc3QgUm9vdCBDQTAeFw0yNDExMjQxMzQwMjFaFw0zNDExMjIxMzQwMjFa
+MGkxCzAJBgNVBAYTAlhYMQ4wDAYDVQQIDAVTdGF0ZTENMAsGA1UEBwwEQ2l0eTEa
+MBgGA1UECgwRVGVzdCBPcmdhbml6YXRpb24xCzAJBgNVBAsMAklUMRIwEAYDVQQD
+DAlsb2NhbGhvc3QwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAwggIKAoICAQCgyHOH
+lzj4eny5Q3o+kce3cIdG3HfxwWemeZtozS7WWq5iZafVK9aDs+68VpYYgMAi1l0h
+Uoj+SZjnlkEf2733drY9eosB/GzJ8mK77mKOuoyaWl0PHtw2gzI1AF7TKrQqErOS
+jMFaZlQl4RtkYsyX2xeF++saww9845exx2VAKVeao3VrDLPOw25yuJlhz5etZbgU
+yqURFmz/Hwybeq+SycX69CFZUEot70ghZvxti3lXFNIOAIRJqPrtqxVzfiLzIpYJ
+oUxpGBVxbNrGHTYLVchzpRKZA3ejUDbRQfkEXlin0r+tmE1QNnMAVFU05gEFayEM
+N2tN9c11s861L0xajIeaf1xi76LJJgozpTl4erfubnngfXSdWjY6G8H/oOSxZxC2
+I1NNvYp78RUiZHdE2EdCY+IZAOniHmD03+3w3YhK+uQR94C/xbFnBCmcnAbzO5Ig
+25X4gCVdG6SXF4bl2UQqO9RHkMRI1ygs+klmNhRC8YV+c/aiN7q50qaFbB35s/DW
+nOFp926sgxlAhoG4ZQLPvEfpT8cIzresTcg8pvSQDrmF1ol8eXEmcqP8O+GZI+9N
+NKKNE9G6raQ8g1p7GQyxTUeKKUo2jyDxWxvyXzjfBdU7cOi9zPIOuMnZajZqlTco
+kKletNhzn/CH9y9427gXlvITWQivRTvDi+ihYQIDAQABo4GXMIGUMB8GA1UdEQQY
+MBaCCWxvY2FsaG9zdIIJMTI3LjAuMC4xMAwGA1UdEwEB/wQCMAAwDgYDVR0PAQH/
+BAQDAgWgMBMGA1UdJQQMMAoGCCsGAQUFBwMBMB0GA1UdDgQWBBQ67nKnNV208ejV
+Dw7wIq9ru4WinjAfBgNVHSMEGDAWgBQp+V1LmqIonZieLtOIRaEfFH3U+zANBgkq
+hkiG9w0BAQsFAAOCAgEAO5pQVtzqIKPasIDE0uLDG8mQ3mwU+M8FMTYK5EF2wVD5
+tPBvg7wOLQIByOHJeW/apQd9uuFKrfduMgDfnQ6d4ztFr1r+ThrIwGj19NBIR8JK
+vtGdpADMV1ygMYv4A6x2gN/2T7Fra2pyL1qc6EtstWs5g/jOweCxTUHCOgIsGFDx
+igU1wm/UVN1fPtSBTwtTPNkSIuyUVAzeqZcH9ZEEi3dqqnx+SN/bzZbRFhe1aSlP
+UPq/2XDlvOUmR3zqFk3t9dhudCtdwUWKUzZEyHvOitQosdQWL3Ywk5vLa63G6oYI
+sVeCgTFdUTBvwQljari5ni8S49G85P/vie+scX4yC/leInNt7THWsFYfgvOeGDfY
+2PgNVlkER7puhuTRDWTybzKyERWPFDveVOIlBLLW2avHCPIOjCGWXzDmWz8/otmQ
+Bq92v5Xzq8NawdvLSacTMFFFRn61Z9zL/XabySjg9+Jj7gLSvTw6iJlFtx9Fl12q
+jDA37+osmDvXR94k33mnZzz7FiuTxrhEjfs0IoFVT3wTUrh00WzcP0fCtegiwfCs
+G9HLyWca/7AT8Dh2Hiwaesuoa/P1kIhWbBNqVQhTh35YanA2a+VjmcO8gL8KzaBy
+yaE+IDxM/2QlkGdr4Vf68eVvuafXksoouDjVCLoMSC1H1UjiuuNWvOrPwCkKqjU=
+-----END CERTIFICATE-----
