+++ cloud/blockstore/tests/recipes/fake-root-kms/certs/server.key
@@ -1,0 +1,52 @@
+-----BEGIN PRIVATE KEY-----
+MIIJQQIBADANBgkqhkiG9w0BAQEFAASCCSswggknAgEAAoICAQCgyHOHlzj4eny5
+Q3o+kce3cIdG3HfxwWemeZtozS7WWq5iZafVK9aDs+68VpYYgMAi1l0hUoj+SZjn
+lkEf2733drY9eosB/GzJ8mK77mKOuoyaWl0PHtw2gzI1AF7TKrQqErOSjMFaZlQl
+4RtkYsyX2xeF++saww9845exx2VAKVeao3VrDLPOw25yuJlhz5etZbgUyqURFmz/
+Hwybeq+SycX69CFZUEot70ghZvxti3lXFNIOAIRJqPrtqxVzfiLzIpYJoUxpGBVx
+bNrGHTYLVchzpRKZA3ejUDbRQfkEXlin0r+tmE1QNnMAVFU05gEFayEMN2tN9c11
+s861L0xajIeaf1xi76LJJgozpTl4erfubnngfXSdWjY6G8H/oOSxZxC2I1NNvYp7
+8RUiZHdE2EdCY+IZAOniHmD03+3w3YhK+uQR94C/xbFnBCmcnAbzO5Ig25X4gCVd
+G6SXF4bl2UQqO9RHkMRI1ygs+klmNhRC8YV+c/aiN7q50qaFbB35s/DWnOFp926s
+gxlAhoG4ZQLPvEfpT8cIzresTcg8pvSQDrmF1ol8eXEmcqP8O+GZI+9NNKKNE9G6
+raQ8g1p7GQyxTUeKKUo2jyDxWxvyXzjfBdU7cOi9zPIOuMnZajZqlTcokKletNhz
+n/CH9y9427gXlvITWQivRTvDi+ihYQIDAQABAoICABhZSEGsRuOTrR1jWM5um6JY
+nMFASbvSBcU3clpvLvIlqSCxSLbm3NAffikwgFTf8/jyogeWAmARZjdgJlGm9gBV
+Y9OF93+KqjD1D5BqpqtQbf2WxLQHAONn1UEjOA5eT7V1apNoKdnrrZT6SlXWrn7V
+c50F9xHw9gCm+Ns37p9GWNl2z/qGYcPkuG5lgqObjAVQO7iOWGwbi3shSR1R+ZJ5
+vbfmTmyXqlGFhKoQ+MEK17U2gSRqCJ6dGWRyaHJlPka4IfpTCTQl17gl0NsizliY
+xPKvzilfqze6F0BzxLyCyzCdy7Wu3w/QMcnxe4viGcfSdElE1UUTqJSexnri2aWO
+uA7NoB+r7fba5oUST6RZhCaWfGtRYMHsEPvYzgxvs8tM3Zzrka6uyMReSjmTyB2e
+yd2yUclC9Qe6n9c3rjaSd8JtbWpZpbC+v3SH+GnksOOPTQL+OWGVVpkXJQUzAIQ3
+wPH6vy42x6yyVQKnGOP3iQdSqHR4SS7/jVU43oF2GIBijqnFu/t6MRrn/mnRwZe8
+Kbiw0A5jefiCIrtEk9RoyZ5Vwd6YsOzM9UE+c/KtfsgHmUuCOJfa80m6XdafEwrK
+9cNXdQFjvVtdQzH/PvHUamfgT4yPQ7MeGqHO274/QTZ2WmFE0C3rpmNNsHDR8C9o
+4FbjMkFdpuTU/OzW1Qg3AoIBAQDKsruX4v137pB0ZGzAs0yGVtOqUY+aUw+Y10qA
+HxLKUbbflUr3wxmJuR53d/BywBA5YQwMXhCzH8nVQjivtZs/C2bs82/ttxmb/q3O
+XNMn0/snuCCEhV8iW72fhQ+Pzq1F6tSqxsM/t3oSHUwSiZY/stQBUN4Q7/2y1PtZ
+O2sju77E9PF7rR618diM3IGphYuQ2oAxYsg73PoT0dUqYGtQiNQ7fg4iyvGmCTs7
+RxOHyVEAb1C8XBe23BRWnmyFMVmBTaAe16odlYDmZVtMUuQtuoWJliaB8b97LfHV
+lhba9AM7syRQcaalYBXG/ZqvpvxmKfTsWgmbuvPpYUTvd5SzAoIBAQDLEBEuKySY
+N8023QGz6hE4Ls3Fs1zqrZo8TF7z5JTg/SnTqHFt2L0JGIoXu4GfEBFpBCV+YQT6
+kQizvvHrt7PSgWy216YSx7VuHq5OOVBcn/NJsuGx9zonzPTfoPnKzLVk0R+l7K2O
+yC8E96DKsy9Ba184gMIWwyrIejt7wn4PObVhP+eugDrhfCVt5ldDE0GuSNljdEEY
+g3Z3TZTiDvrJCzkDv+bF+K+rzElEr1AS7nGHLj4cTNp2DIzVmr8luWmpKy7Y53Wn
+DCk9CZa5weDk737Ys8Se0zu2wodehGJ5H/5NaNbwXHKflwCYpacHUszU07zaSq0h
+ANG4iYTl7oObAoIBAD4L2gqTTB1TJFrMhtGx0MHqgB5JmxiVRlf1xvNTD7BotP7Z
+tzoCOaYDyYSubSbjDdRSlZqPjLMYDvM9sxZuE9iHh/kOR3vPqYIrwk5B8czmQaIV
+5bZLEST2SiQzWeEeVcHqMTEZaJjvc+lpmGiwPxK+jdDdTSyIGfOM6GeXHaOxi6LE
+i8rc4mK1ZiEwOmyG2YL8r3LulvZkHRTGDWdMXm0nhTe0zWTvFGtLthreFedhw27L
+pD4UPpoV4g/lrLhUs941wcbPI16s7ekYq4rAcOJijw/fq2fpdUCifr1vlQ1u1jKH
+Ep2hPG6w69metaQBDI/FVhinMvu9jX4p7S0UQKsCggEACgLVxf/sgvClVmIY9mTA
+ypNMI8YO+6bpZrWhLMcFK/5vNHcmeHLqBuvFyxu8bToekZ3Zt1xTIYNK+YmdZbCd
+1e/5ix4gmuZrpG0wzLx4oMhl9ykxhVOZ82JgZDIMfpoPC1RgCg8cE2D9KHKI49Vn
+rAL+I8v9RaADmPSlwBk7m8aNkajT5D+AJ6O2OK/Vd4IYh94hkYymtO2uRn11rJtu
+c8Dw46VvXC1Bg+fDaPZ3rSdCQ5bdCwoi7xT5UIKbj+GPOv7dBWaL31tzIQY6/xd2
+3fKMv2RFeksRs5XtlJRUQoJ4SBq2Fg2IvSN1rRDlpi/H12pE8kIiPPFMZZuqCKf0
+pwKCAQAJ+U8wv8i8gl9RMHnbpoBw7sL6HfhhBj/zfjJ1s7Y34gnDvzZRxnqcq0y3
+6AVW+rmw3CSb+EYFObyUrRAKowwMsOngfsVBTN8nvIqnMGOy379oVDn+2YNBjrwk
+dZ371SedC+7lf1urB0r5tmdzegLslaO4Kx3oW1bgHfu2uRMwmBpP1+BQ8UO0fz7W
+KQP8cxsodY2IRRCx06pN16ZyXnWF4xkpnBM//AahaluZDpGPksQLggO3ssCRGezf
+Y20Fyqe85OrMnLmywWYr+ivY16ZyP2Jot8nMstFQ4NwbUfgCI5IqmZU1RyLwoL4l
+atLx7RMyik0OywYElyYZrGkgXomd
+-----END PRIVATE KEY-----
