+++ cloud/blockstore/tests/recipes/fake-root-kms/ya.make
@@ -1,0 +1,13 @@
+PY3_PROGRAM()
+
+PY_SRCS(
+    __main__.py
+)
+
+PEERDIR(
+    cloud/blockstore/tests/python/lib
+    cloud/tasks/test/common
+    library/python/testing/recipe
+)
+
+END()
