+++ cloud/blockstore/tests/recipes/ya.make
@@ -1,5 +1,6 @@
 RECURSE(
     endpoint
+    fake-root-kms
     local-kikimr
     local-null
     notify-mock
